requirements.txt
astroid==3.3.11
azure-core==1.35.1
azure-storage-blob==12.26.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.2
dill==0.4.0
dotenv==0.9.9
idna==3.10
isodate==0.7.2
isort==6.1.0
mccabe==0.7.0
numpy==2.3.3
pandas==2.3.3
platformdirs==4.5.0
pyarrow==21.0.0
pycparser==2.23
pylint==3.3.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
RapidFuzz==3.14.1
requests==2.32.5
ruff==0.14.0
six==1.17.0
tomlkit==0.13.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
